store raw adc/battery readings

+++ Functions/Si7021_Reading.cs
@@ -5,7 +5,7 @@ namespace Karmatach.MaxPlay
         public string Time { get; set; }
         public double F { get; set; }
         public double RH { get; set; }
-        public double Battery { get; set; }
+        public ushort Battery { get; set; }
         public bool? SetHeaterStateTo { get; set; }
     }
 }
